requirements.txt
paho-mqtt>=1.6.1
pymodbus>=3.0.0
pyyaml>=6.0
numpy>=1.22.0
matplotlib>=3.5.0
loguru>=0.6.0
